adding element-values function
Adding various flavors of element-values function.  It will attempt to
mimic the behavior of cts:element-values().

I’m not sure if we have a real way to support the $quality-weight with
triples.  I also attempted to mimic the $start behavior, but it’s not
well tested.

One major difference is that the resulting values cannot be used
natively by cts:frequency().